volume/mounts: split tests per parser
